Move PaddingValues.plus to compose-ui module
This allows the operator to also be used by the design system internal
module itself.
